requirements.txt
numpy==1.25.2
pandas==2.0.3
pytest==7.4.4
torch==2.2.1
transformers==4.38.2
pyahocorasick